+++ recommendation/src/main/kotlin/com/mconsulting/indexrecommender/IndexCoalesceEngine.kt
@@ -16,6 +16,7 @@ class IndexCoalesceEngine {
 
     fun coalesce(candidateIndexes: List<Index>): IndexCoalesceResult {
         if (candidateIndexes.isEmpty()) return IndexCoalesceResult(candidateIndexes)
+
         // Final indexes
         val indexes = mutableListOf<Index>()
         val removedIndexes = mutableListOf<Index>()
@@ -48,6 +49,8 @@ class IndexCoalesceEngine {
     }
 
     private fun processTextIndexes(textIndexes: List<TextIndex>, compoundTextIndexes: List<CompoundTextIndex>, removedIndexes: MutableList<Index>, indexes: MutableList<Index>) {
+        val shapes = mutableListOf<ShapeStatistics>()
+
         // Do we have any text indexes
         if (textIndexes.isNotEmpty()) {
             val textIndexQueue = LinkedList(textIndexes.sortedBy { it.fields.size })
@@ -57,42 +60,51 @@ class IndexCoalesceEngine {
             textIndexQueue.forEach {
                 mergedFields += it.fields.filter { !mergedFields.contains(it) }
                 removedIndexes += it
+                shapes += it.statistics
             }
 
             // Go over all the compound text indexes
             if (compoundTextIndexes.isNotEmpty()) {
-                mergedFields += mergeCompoundIndexes(compoundTextIndexes, indexes)
+                mergedFields += mergeCompoundIndexes(compoundTextIndexes, indexes, shapes)
             }
 
             // Merge any compound text indexes, splitting them and merging them
-            indexes += TextIndex(createIndexName(mergedFields.toList()), mergedFields.toList())
+            val index = TextIndex(createIndexName(mergedFields.toList()), mergedFields.toList())
+            index.statistics = shapes
+            indexes += index
             return
         }
 
         // Do we have more than one compound index, merge them
         if (compoundTextIndexes.size > 1) {
-            val mergedFields = mergeCompoundIndexes(compoundTextIndexes, indexes)
-            indexes += TextIndex(createIndexName(mergedFields.toList()), mergedFields.toList())
+            val mergedFields = mergeCompoundIndexes(compoundTextIndexes, indexes, shapes)
+            val index = TextIndex(createIndexName(mergedFields.toList()), mergedFields.toList())
+            index.statistics = shapes
+            indexes += index
         }
     }
 
-    private fun mergeCompoundIndexes(compoundTextIndexes: List<CompoundTextIndex>, indexes: MutableList<Index>): List<TextField> {
+    private fun mergeCompoundIndexes(compoundTextIndexes: List<CompoundTextIndex>, indexes: MutableList<Index>, shapes: MutableList<ShapeStatistics>): List<TextField> {
         val mergedFields = mutableListOf<TextField>()
         // We need to merge any missing text fields
         mergedFields += compoundTextIndexes.map {
             it.fields.filter { !mergedFields.contains(it) }
         }.flatten()
 
-        compoundTextIndexes.forEach {
-            val filteredFields = it.compoundFields.filter {
-                it.name !in listOf("_fts", "_ftsx")
+        compoundTextIndexes.forEach { textIndex ->
+            val filteredFields = textIndex.compoundFields.filter { field ->
+                field.name !in listOf("_fts", "_ftsx")
             }
+
+            // Add any shapes
+            shapes += textIndex.statistics
+
             // Create a candidate compound index and add it if it does
             // not already exist, we are in fact splitting indexes here
             // due to there only being possible to have one text index per collection
-            val compoundIndex = CompoundIndex(filteredFields.map {
-                "${it.name}_${when (it.direction) {
-                    IndexDirection.ASCENDING, IndexDirection.DESCENDING -> it.direction.value()
+            val compoundIndex = CompoundIndex(filteredFields.map { field ->
+                "${field.name}_${when (field.direction) {
+                    IndexDirection.ASCENDING, IndexDirection.DESCENDING -> field.direction.value()
                     else -> 1
                 }}"
             }.joinToString("_"), filteredFields)
@@ -118,7 +130,19 @@ class IndexCoalesceEngine {
         // Add any indexes that are not covered by a composite index
         indexes += singleIndexes.filter { singleFieldIndex ->
             compoundFieldIndexes.firstOrNull { multiFieldIndex ->
-                multiFieldIndex.fields.first().name == singleFieldIndex.name
+
+                // Merge any statistics when we are removing fields
+                if (multiFieldIndex.fields.first().name == singleFieldIndex.name) {
+                    singleFieldIndex.statistics.forEach { shape ->
+                        if (!multiFieldIndex.statistics.contains(shape)) {
+                            multiFieldIndex.statistics += shape
+                        } else {
+                            multiFieldIndex.statistics.find { it == shape }!!.merge(shape)
+                        }
+                    }
+                }
+
+                multiFieldIndex.fields.first().name == singleFieldIndex.field.name && !singleFieldIndex.unique
             } == null
         }
 
@@ -145,6 +169,14 @@ class IndexCoalesceEngine {
                 if (index != otherIndex && index.fields.size <= otherIndex.fields.size) {
                     if (index.fields == otherIndex.fields.subList(0, index.fields.size) && !index.unique) {
                         removedIndexes += index
+                        // Merge statistics
+                        index.statistics.forEach { shape ->
+                            if (!otherIndex.statistics.contains(shape)) {
+                                otherIndex.statistics += shape
+                            } else {
+                                otherIndex.statistics.find { it == shape }!!.merge(shape)
+                            }
+                        }
                     }
                 }
             }