+++ docs_template/content/HISTORY.md
@@ -1,4 +1,6 @@
 1.0.0
 =====
-- Basic extract Schema extraction functionality.
-- Basic apply Schema functionality.
+- Extract existing Indexes from MongoDB collections
+- Process `system.profile` collection and extract candidate indexes.
+- Process MongoDB logs and extract candidate indexes.
+- Coalesce all candidate indexes into a minimal list of optimized indexes.